refactor(directory structure): move files to new directories
Moved the files in the src/containers/ folder to src/pages and src/containers/layout to src/layout

fix(auth): refresh token issue

Add missing refresh token logic and fix the issue when user is logged out after a certain period of time.

feat(goal): add detail view

Add a new feature to allow users to view the details of each goal when clicking on it.